Optimize deployment
