+++ bin/langfuse-ecr-ecs-deployment-cdk.ts
@@ -2,24 +2,45 @@
 import 'source-map-support/register';
 
 import * as cdk from 'aws-cdk-lib';
-import { LangfuseEcrEcsDeploymentCdkStack } from '../lib/langfuse-ecr-ecs-deployment-cdk-stack';
+import * as dotenv from 'dotenv';
+import { CdkLangfuseEcrEcsDeploymentStack } from '../lib/langfuse-ecr-ecs-deployment-cdk-stack';
+import { checkEnvVariables } from '../utils/check-environment-variable';
+import { IEnvTypes } from '../process-env-typed';
 
+dotenv.config(); // Load environment variables from .env file
 const app = new cdk.App();
+
+const { CDK_DEFAULT_ACCOUNT: account, CDK_DEFAULT_REGION: region } = process.env;
+
+const cdkRegions = process.env.CDK_DEPLOY_REGIONS?.split(',') ?? [region]; // Parsing comma separated list of regions
+const deployEnvironments = process.env.ENVIRONMENTS?.split(',') ?? ['dev']; // Parsing comma separated list of environments
 
 export const LATEST_IMAGE_VERSION = 'latest';
 
-new LangfuseEcrEcsDeploymentCdkStack(app, 'LangfuseEcrEcsDeploymentCdkStack', {
-  /* If you don't specify 'env', this stack will be environment-agnostic.
-   * Account/Region-dependent features and context lookups will not work,
-   * but a single synthesized template can be deployed anywhere. */
+// check general stack props
+checkEnvVariables('ECR_REPOSITORY_NAME', 'APP_NAME', 'IMAGE_VERSION', 'PORT');
 
-  /* Uncomment the next line to specialize this stack for the AWS Account
-   * and Region that are implied by the current CLI configuration. */
-  // env: { account: process.env.CDK_DEFAULT_ACCOUNT, region: process.env.CDK_DEFAULT_REGION },
+const envTyped: IEnvTypes = {
+    ECR_REPOSITORY_NAME: process.env.ECR_REPOSITORY_NAME!,
+    APP_NAME: process.env.APP_NAME!,
+    IMAGE_VERSION: process.env.IMAGE_VERSION!,
+    PORT: process.env.PORT!,
+};
 
-  /* Uncomment the next line if you know exactly what Account and Region you
-   * want to deploy the stack to. */
-  // env: { account: '123456789012', region: 'us-east-1' },
-
-  /* For more information, see https://docs.aws.amazon.com/cdk/latest/guide/environments.html */
-});
+for (const cdkRegion of cdkRegions) {
+  for (const environment of deployEnvironments) {
+    new CdkLangfuseEcrEcsDeploymentStack(app, 'CdkLangfuseEcrEcsDeploymentStack', {
+      env: {
+        account,
+        region: cdkRegion,
+      },
+      tags: {
+        environment,
+      },
+      deployRegion: cdkRegion,
+      environment,
+      envTyped,
+      description: `Langfuse ECR/ECS deployment stack for ${environment} environment in ${cdkRegion} region.`,
+    });
+  }
+}
